refactor: remove GetEarlyAccessButton
